github: Handle commit_comment event
Review issues were added another section - review blockers. In case
there is !blocker string in a commit comment, a new review blocker item
is added to the blocker checklist for the given review issue. In this
way the reviewer can keep track of what needs to be fixed, and the
commit author as well.

The blocker command must be on a standalone line and what is following
the command until the end of the line is taken as the blocker summary.

Change-Id: 9434df3884
Story-Id: SF-180