utils/log: Make logging more configurable
The issue was that the package was logging the function that was calling the
logging functions, but there are cases when that is a useless information and
you actually need to configure what function in the stack to mention.

There is Logger struct now that keeps the number of callers to skip and the
number can be increased and decreased. There are also some global functions just
using the default settings.

Story-Id: unassigned
Change-Id: 1fb0df626e