Refactor the GitHub utility package
There is now a server that can be used to handle GitHub webhooks. All
the package user needs to do is to supply an object implementing the
right interface so that the event is handled, e.g. in case there is
HandleIssuesEvent method, it will be used to handle issues_event coming
from GitHub.

Change-Id: 614c2e5706
Story-Id: salsaflow/salsaflow-daemon#9