Pivotal Tracker: Allow for parallel review and QA
The activity handler for Pivotal Tracker was changed so that it does not remove
QA state labels when the associated story review request is reopened or closed.
This basically removes the restriction that QA needs to happen strictly after
review. On the other hand, this makes it possible to add code without anybody
testing it properly. This is a known limitation.

Change-Id: ce59167044
Story-Id: salsaflow/salsaflow-daemon#13